Fixed basic api
